+++ config/test.exs
@@ -7,5 +7,6 @@ config :tds,
     hostname: System.get_env("SQL_HOSTNAME") || "127.0.0.1",
     username: System.get_env("SQL_USERNAME") || "sa",
     password: System.get_env("SQL_PASSWORD") || "some!Password",
-    database: "test"
+    database: "test",
+    set_allow_snapshot_isolation: :on
   ]